+++ spec/lib/examples/simple_schema_spec.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 LineItem = Scheming.object do
-  attribute :id, Integer
+  attribute :id, Union(Integer, String)
   attribute :name, String
   attribute :taxable, :bool
   attribute :price, Float
@@ -37,7 +37,12 @@ RSpec.describe do
             additionalProperties: false,
             required: %w[id name taxable price],
             properties: {
-              id: { type: 'integer' },
+              id: {
+                oneOf: [
+                  { type: 'integer' },
+                  { type: 'string' }
+                ]
+              },
               name: { type: 'string' },
               taxable: { type: 'boolean' },
               desc: {